What was turned in
The updated DBMS_program.cpp on this push is the one I uploaded to
csnet. Was trying to get the program to take in user inputs for more
user interaction, but that wasn't working so everything is on rails.
Turned it in at about 3:40 so I'm guessing that like minus 4 pts. Was
trying to do more but wasn't working out.